% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OECD_misc.R
\name{oecd_misc}
\alias{oecd_misc}
\alias{oecd_members}
\alias{oecd_member_metrics}
\alias{oecd_secretary}
\title{Scraping tools for OECD related information.}
\source{
https://en.wikipedia.org/wiki/OECD
}
\usage{
oecd_members(code = FALSE)

oecd_member_metrics()

oecd_secretary()
}
\arguments{
\item{code}{Logical. If \code{TRUE}, prints the code to create the vector of
OECD countries to the console. \code{FALSE} by default, returns the whole
dataframe.}
}
\description{
Scraping tools for OECD related information.

\code{oecd_members()}Scrapes member information of OECD countries from
Wikipedia and returns it in the form of a dataframe. This dataframe contains
member information such as accession dates, country names, etc.

\code{oecd_member_metrics()} Scrapes member information of OECD countries from
Wikipedia and returns it in the form of a dataframe. This dataframe contains
economic data of member countries. Not to be used for analysis purposes.

\code{oecd_secretary()} scrapes the table of previous OECD Secretary-Generals
from Wikipedia and returns a dataframe with their details.
}
\details{
This family of functions returns information related to the OECD
such as the member States, the name and tenure period of the secretary-
generals of the OECD, etc. Data should be used carefully and not for
analytical purposes.
}
\examples{
library(magrittr)
oecd_members()
library(magrittr)
oecd_member_metrics()
library(magrittr)
oecd_secretary()
}
