% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OECD_palette.R
\docType{data}
\name{oecd_palette}
\alias{oecd_palette}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
oecd_palette
}
\description{
Discrete color palette based on OECD colors.
}
\keyword{datasets}
