% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OECD_palette.R
\docType{data}
\name{oecd_palettes}
\alias{oecd_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 1.
}
\usage{
oecd_palettes
}
\description{
Discrete color palette based on OECD colors.
}
\keyword{datasets}
