% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OECD_palette.R
\name{oecd_palette}
\alias{oecd_palette}
\title{An OECD palette generator}
\source{
Adapted from https://github.com/karthik/wesanderson/blob/master/R/colors.R
}
\usage{
oecd_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Current choice is:
\code{oecd}}

\item{n}{Number of colors desired. If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
This is a small palette generator for OECD graphs that generates color vectors
to be used in graphs
}
\examples{
oecd_palette("oecd")

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- oecd_palette(21, name = "Centres", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
