% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WP6RTools-package.R
\docType{package}
\name{WP6RTools-package}
\alias{WP6RTools}
\alias{WP6RTools-package}
\title{WP6RTools: R Tools for WP6 Data Analysis Work}
\description{
R Tools for WP6 Data Analysis Work.
}
\author{
\strong{Maintainer}: Bernhard Bieri \email{bernhard.bieri@graduateinstitute.ch} (\href{https://orcid.org/0000-0001-5943-9059}{ORCID})

}
\keyword{internal}
